requirements.txt
Django==3.1.2
djangorestframework==3.11.1
flake8==3.8.4
psycopg2-binary==2.8.6
python-decouple==3.3
drf-yasg==1.17.0